sentiment-grpc-service/requirements.txt
transformers
torch
grpcio
grpcio-tools
protobuf
python-dotenv